[Android] Extract SSL key manager factory

Copilot-Session: d4c4e834-d70e-482c-88aa-d762e1d390bd
