+++ src/libraries/System.Net.Security/src/System/Net/Security/Pal.Android/SafeDeleteSslContext.cs
@@ -6,11 +6,8 @@ using System.Collections.Generic;
 using System.Net.Security;
 using System.Runtime.InteropServices;
 using System.Security.Authentication;
-using System.Security.Cryptography;
-using System.Security.Cryptography.X509Certificates;
 using System.Threading;
 
-using PAL_KeyAlgorithm = Interop.AndroidCrypto.PAL_KeyAlgorithm;
 using PAL_SSLStreamStatus = Interop.AndroidCrypto.PAL_SSLStreamStatus;
 
 namespace System.Net
@@ -221,7 +218,7 @@ namespace System.Net
             IntPtr keyManagers = IntPtr.Zero;
             if (authOptions.CertificateContext is not null)
             {
-                keyManagers = CreateKeyManagers(authOptions.CertificateContext);
+                keyManagers = AndroidKeyManager.Create(authOptions.CertificateContext);
             }
             else if (!authOptions.IsServer && authOptions.CertSelectionDelegate is not null)
             {
@@ -252,73 +249,6 @@ namespace System.Net
                     Interop.JObjectLifetime.DeleteGlobalReference(keyManagers);
                 }
             }
-        }
-
-        internal static IntPtr CreateKeyManagers(SslStreamCertificateContext context)
-        {
-            X509Certificate2 cert = context.TargetCertificate;
-            Debug.Assert(cert.HasPrivateKey);
-
-            IntPtr keyManagers;
-            if (Interop.AndroidCrypto.IsKeyStorePrivateKeyEntry(cert.Handle))
-            {
-                keyManagers = Interop.AndroidCrypto.SSLStreamCreateKeyManagersFromKeyStoreEntry(cert.Handle);
-            }
-            else
-            {
-                PAL_KeyAlgorithm algorithm;
-                byte[] keyBytes;
-                using (AsymmetricAlgorithm key = GetPrivateKeyAlgorithm(cert, out algorithm))
-                {
-                    keyBytes = key.ExportPkcs8PrivateKey();
-                }
-
-                try
-                {
-                    IntPtr[] ptrs = new IntPtr[context.IntermediateCertificates.Count + 1];
-                    ptrs[0] = cert.Handle;
-                    for (int i = 0; i < context.IntermediateCertificates.Count; i++)
-                    {
-                        ptrs[i + 1] = context.IntermediateCertificates[i].Handle;
-                    }
-
-                    keyManagers = Interop.AndroidCrypto.SSLStreamCreateKeyManagers(keyBytes, algorithm, ptrs);
-                }
-                finally
-                {
-                    CryptographicOperations.ZeroMemory(keyBytes);
-                }
-            }
-
-            if (keyManagers == IntPtr.Zero)
-            {
-                throw new Interop.AndroidCrypto.SslException();
-            }
-
-            return keyManagers;
-        }
-
-        private static AsymmetricAlgorithm GetPrivateKeyAlgorithm(X509Certificate2 cert, out PAL_KeyAlgorithm algorithm)
-        {
-            AsymmetricAlgorithm? key = cert.GetRSAPrivateKey();
-            if (key != null)
-            {
-                algorithm = PAL_KeyAlgorithm.RSA;
-                return key;
-            }
-            key = cert.GetECDsaPrivateKey();
-            if (key != null)
-            {
-                algorithm = PAL_KeyAlgorithm.EC;
-                return key;
-            }
-            key = cert.GetDSAPrivateKey();
-            if (key != null)
-            {
-                algorithm = PAL_KeyAlgorithm.DSA;
-                return key;
-            }
-            throw new NotSupportedException(SR.net_ssl_io_no_server_cert);
         }
 
         private unsafe void InitializeSslContext(
